+++ ChangeLog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added function composition
 - added partial functions
+- added functions partition, distinct, and dedupe
 
 ### Removed